pyproject.toml
[tool.poetry]
name = "web-youtube-dl"
version = "0.0.0"
description = "A web version of youtube-dl"
readme = "README.rst"
authors = ["Uriel Mandujano <uriel.mandujano14@gmail.com>"]
classifiers = [
  "Topic :: Home Automation",
  "Topic :: Multimedia :: Sound/Audio",
  "Topic :: Multimedia :: Sound/Audio :: Capture/Recording",
]

[tool.poetry.scripts]
web-youtube-dl-cli = "web_youtube_dl.app.main:cli_download"
web-youtube-dl-web = "web_youtube_dl.app.main:run_app"

[build-system]
requires = ["poetry-core>=1.0.0", "poetry-dynamic-versioning>=1.0.0,<2.0.0"]
build-backend = "poetry_dynamic_versioning.backend"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0"
uvicorn = "^0.17.0"
python-multipart = "^0.0.5"
aiofiles = "^0.5.0"
janus = "^0.5.0"
Werkzeug = "^1.0.1"
pytube = "^15.0.0"
Jinja2 = "^3.1.2"
websockets = "^10.3"
ffmpeg-python = "^0.2.0"
musicbrainzngs = "^0.7.1"
mutagen = "^1.45.1"
pydantic = "^2.4.2"
sqlalchemy = { extras = ["asyncio"], version = "^2.0.23" }
arrow = "^1.3.0"
sqlalchemy-utils = "^0.41.1"
aiosqlite = "^0.19.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0"
pytest-cov = "^2.10.1"
types-setuptools = "^57.4.17"
types-requests = "^2.27.30"
requests = "^2.28.0"
ruff = "^0.2.2"


[tool.poetry.group.server.dependencies]
sqlalchemy = "^2.0.25"
sqlalchemy-utils = "^0.41.1"
arrow = "^1.3.0"
sqlmodel = "^0.0.16"

[tool.ruff]
select = [
  "E",   # pycodestyle
  "F",   # pyflakes
  "UP",  # pyupgrade
  "ERA", # eradicate
]

[tool.ruff.isort]
known-first-party = ["web_youtube_dl", "tests"]

[tool.poetry-dynamic-versioning]
enable = true
metadata = true
pattern = "default-unprefixed"
